Address PR #3 code review findings
- Restore blank lines unrelated to the MUI migration that @mui/codemod's
  recast reprint dropped in Expenses.tsx, Categories.tsx, Settings.tsx.
- Convert double-quoted strings in codemod-generated sx blocks to single
  quotes per .prettierrc, then run prettier on the now-fully-exposed
  formatting drift in Login/Register/Dashboard/NotFound/Settings.tsx
  (verified develop's versions were already prettier-clean, so --write
  only touched the codemod-introduced regions).
- Extract the duplicated sx={{ color: 'text.secondary' }} into a shared
  secondaryTextSx constant (helpers/sxHelpers.ts).
- Extract ExpenseSchema out of ExpenseForm.tsx into its own
  expenseSchema.ts so ExpenseForm.test.ts no longer transitively imports
  MUI/x-date-pickers just to test a Yup schema.
- Remove vitest.config.ts's server.deps.inline workaround entirely, now
  that no test file imports MUI -- it's no longer needed at all rather
  than just generalized. Test run time dropped from ~8s to ~0.7s.
