+++ apps/web/src/helpers/sxHelpers.ts
@@ -1,0 +1,1 @@
+export const secondaryTextSx = { color: 'text.secondary' };
