+++ apps/web/src/pages/Categories.tsx
@@ -31,6 +31,7 @@ import {
     ColorLens as ColorLensIcon
 } from '@mui/icons-material';
 import { useCategoryStore } from '@/store/categoryStore';
+import { secondaryTextSx } from '@/helpers/sxHelpers';
 import { Category } from '@expenses/shared';
 
 export default function Categories() {
@@ -178,15 +179,14 @@ export default function Categories() {
                     Add Category
                 </Button>
             </Box>
+
             {loading ? (
                 <Box sx={{ display: 'flex', justifyContent: 'center', my: 4 }}>
                     <CircularProgress />
                 </Box>
             ) : categories.length === 0 ? (
                 <Paper sx={{ p: 3, textAlign: 'center' }}>
-                    <Typography variant="body1" sx={{
-                        color: "text.secondary"
-                    }}>
+                    <Typography variant="body1" sx={secondaryTextSx}>
                         No categories found. Add your first category to get started!
                     </Typography>
                 </Paper>
@@ -219,9 +219,7 @@ export default function Categories() {
                                                 mr: 1
                                             }}
                                         />
-                                        <Typography variant="body2" sx={{
-                                            color: "text.secondary"
-                                        }}>
+                                        <Typography variant="body2" sx={secondaryTextSx}>
                                             {category.color || 'No color set'}
                                         </Typography>
                                     </Box>
@@ -249,6 +247,7 @@ export default function Categories() {
                     ))}
                 </Grid>
             )}
+
             {/* Add/Edit Category Dialog */}
             <Dialog open={dialogOpen} onClose={handleDialogClose}>
                 <DialogTitle>{dialogMode === 'add' ? 'Add Category' : 'Edit Category'}</DialogTitle>
@@ -304,6 +303,7 @@ export default function Categories() {
                     </Button>
                 </DialogActions>
             </Dialog>
+
             {/* Delete Confirmation Dialog */}
             <Dialog open={deleteDialogOpen} onClose={() => setDeleteDialogOpen(false)}>
                 <DialogTitle>Confirm Delete</DialogTitle>
@@ -324,6 +324,7 @@ export default function Categories() {
                     </Button>
                 </DialogActions>
             </Dialog>
+
             {/* Notification Snackbar */}
             <Snackbar
                 open={notification.open}
