+++ apps/web/src/pages/Expenses.tsx
@@ -29,6 +29,7 @@ import ExpenseTable from '../components/ExpenseTable';
 import ExpenseForm from '../components/ExpenseForm';
 import { getExpenses, deleteExpense, createExpense, updateExpense } from '../services/expenseService';
 import { useCategoryStore } from '@/store/categoryStore';
+import { secondaryTextSx } from '@/helpers/sxHelpers';
 import { Expense, ExpenseCreateInput } from '@expenses/shared';
 
 export default function Expenses() {
@@ -201,6 +202,7 @@ export default function Expenses() {
                     Add Expense
                 </Button>
             </Box>
+
             <Paper sx={{ p: 2, mb: 3 }}>
                 <Box sx={{ display: 'flex', justifyContent: 'space-between', alignItems: 'center', mb: 2 }}>
                     <TextField
@@ -270,15 +272,14 @@ export default function Expenses() {
                     </Grid>
                 )}
             </Paper>
+
             {loading ? (
                 <Box sx={{ display: 'flex', justifyContent: 'center', my: 4 }}>
                     <CircularProgress />
                 </Box>
             ) : expenses.length === 0 ? (
                 <Paper sx={{ p: 3, textAlign: 'center' }}>
-                    <Typography variant="body1" sx={{
-                        color: "text.secondary"
-                    }}>
+                    <Typography variant="body1" sx={secondaryTextSx}>
                         No expenses found. Add your first expense to get started!
                     </Typography>
                 </Paper>
@@ -301,6 +302,7 @@ export default function Expenses() {
                     </Box>
                 </>
             )}
+
             {/* Add/Edit Expense Form Dialog */}
             <ExpenseForm
                 open={formOpen}
@@ -310,6 +312,7 @@ export default function Expenses() {
                 initialValues={editExpense || undefined}
                 title={editExpense ? 'Edit Expense' : 'Add Expense'}
             />
+
             {/* Delete Confirmation Dialog */}
             <Dialog open={deleteDialogOpen} onClose={() => setDeleteDialogOpen(false)}>
                 <DialogTitle>Confirm Delete</DialogTitle>
@@ -326,6 +329,7 @@ export default function Expenses() {
                     </Button>
                 </DialogActions>
             </Dialog>
+
             {/* Notification Snackbar */}
             <Snackbar
                 open={notification.open}
